+++ index.js
@@ -34,8 +34,8 @@ class getNodeLTS {
         for (const v of versions) {
           const major = semver.major(v.version)  // ex: v12, v10, ...
 
-          if (v.lts === false) continue      // ignore all but LTS
-          // if (major % 2 !== 0) continue
+          // if (v.lts === false) continue      // ignore all but LTS
+          if (major % 2 !== 0) continue         // ignore odd releases
 
           // find the earliest LTS release for each major
           if (!this.majorsInitial[major]) this.majorsInitial[major] = v
@@ -66,13 +66,16 @@ class getNodeLTS {
 
   getActive (opts) {
     const now = new Date().getTime()
-    return Object.keys(this.filter(this.majorsLatest, ([maj, obj]) => {
+    function filterActive ([maj, obj]) {
       return new Date(obj.dateEndLTS).getTime() > now
-    }))
+    }
+    function filterActiveLTS ([maj, obj]) {
+      return obj.lts !== false && new Date(obj.dateEndLTS).getTime() > now
+    }
+    return Object.keys(this.filter(this.majorsLatest, opts.lts ? filterActiveLTS : filterActive))
   }
 
   json (opts = {}) {
-    // console.log(JSON.stringify(versions))
     return JSON.stringify(this.getActive(opts))
   }