+++ main.js
@@ -1,4 +1,4 @@
 const l=require('./index')
 l.fetch().then(() => {
-  console.log(`::set-output name=lts::${l.json()}`)
+  console.log(`::set-output name=active::${l.json()} name=lts::${l.json({lts: true})}`)
 })