+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ### 1.2.0 - 2022-05-31
 
+- fix: output was invalid
 - drop @actions/core dependency
 
 
 ### 1.1.0 - 2021-02-21
 
 - switch default output from YAML to JSON
+- setOutput